Move penv setup out of main.py / invert mcu detection logic (#255)
